improve database support and improved input validation

+++ command_handlers.py
@@ -1,6 +1,6 @@
 import telegram.ext
 import logging
-from util import get_arguments, clean_input
+from util import process_input
 from data_model import ItemList, Item, db_session
 
 
@@ -21,11 +21,9 @@ def rick(update: telegram.Update, context: telegram.ext.CallbackContext):
 
 
 def new(update: telegram.Update, context: telegram.ext.CallbackContext):
-    list_name = update.message.text.lstrip('/new')
-    list_name = clean_input(list_name)
+    list_name, args = process_input(update.message.text)
     response = ''
-
-    if len(list_name) > 0:
+    try:
         session = db_session()
         existing_list = session.query(ItemList).filter(ItemList.list_name == list_name).all()
         if len(existing_list) > 0:
@@ -36,7 +34,7 @@ def new(update: telegram.Update, context: telegram.ext.CallbackContext):
             session.commit()
             response += '"' + list_name + '" is aangemaakt!'
         session.close()
-    else:
+    except ValueError:
         response += 'Lullo, je moet wel wat invullen he'
     context.bot.send_message(chat_id=update.effective_chat.id, text=response)
 
@@ -44,9 +42,9 @@ def new(update: telegram.Update, context: telegram.ext.CallbackContext):
 def pleepapier(update: telegram.Update, context: telegram.ext.CallbackContext):
     list_name = 'Pleepapier'
     session = db_session()
-    item_list = session.query(ItemList).order_by(ItemList.list_name == list_name).first()
+    item_list = session.query(Item).filter(Item.item_list == list_name).order_by(Item.id).all()
     pleepapier_string = 'Pleepapier:\n'
-    items = [str(index + 1) + '. ' + item.item_name + ' (' + item.created_by + ')'for index, item in enumerate(item_list.items)]
+    items = [str(index + 1) + '. ' + item.item_name + ' (' + item.created_by + ')\n'for index, item in enumerate(item_list)]
     pleepapier_string = pleepapier_string + ''.join(items)
     context.bot.send_message(chat_id=update.effective_chat.id, text=pleepapier_string)
     session.close()
@@ -55,9 +53,9 @@ def pleepapier(update: telegram.Update, context: telegram.ext.CallbackContext):
 def reservelijst(update: telegram.Update, context: telegram.ext.CallbackContext):
     list_name = 'Reservelijst'
     session = db_session()
-    item_list = session.query(ItemList).order_by(ItemList.list_name == list_name).first()
+    item_list = session.query(Item).filter(Item.item_list == list_name).order_by(Item.id).all()
     reservelijst_string = 'Reservelijst:\n'
-    items = [str(index + 1) + '. ' + item.item_name + ' (' + item.created_by + ')' for index, item in enumerate(item_list.items)]
+    items = [str(index + 1) + '. ' + item.item_name + ' (' + item.created_by + ')\n' for index, item in enumerate(item_list)]
     reservelijst_string = reservelijst_string + ''.join(items)
     context.bot.send_message(chat_id=update.effective_chat.id, text=reservelijst_string)
     session.close()
@@ -67,14 +65,10 @@ def add(update: telegram.Update, context: telegram.ext.CallbackContext):
     list_name = 'Pleepapier'
     added_message = ''
     try:
-        command = update.message.text.lstrip('/add')
-        arguments = get_arguments(command)
-        if arguments:
-            if '--reserve' in arguments or '-r' in arguments:
+        item_to_add, args = process_input(update.message.text)
+        if args:
+            if '--reserve' in args or '-r' in args:
                 list_name = 'Reservelijst'
-        item_to_add = clean_input(command, arguments)
-        if item_to_add is None or len(item_to_add) == 0:
-            raise ValueError('No input given')
         session = db_session()
         item = Item(item_name=item_to_add, item_list=list_name, created_by=update.effective_user.first_name)
         session.add(item)
@@ -91,24 +85,31 @@ def remove(update: telegram.Update, context: telegram.ext.CallbackContext):
     list_name = 'Pleepapier'
     removed_message = ''
     try:
-        command = update.message.text.lstrip('/remove')
-        arguments = get_arguments(command)
-        if arguments:
-            if '--reserve' in arguments or '-r' in arguments:
+        command, args = process_input(update.message.text)
+        if args:
+            if '--reserve' in args or '-r' in args:
                 list_name = 'Reservelijst'
-        command = clean_input(command, arguments)
         session = db_session()
-        item = session.query(Item).filter(Item.item_name == command, Item.item_list == list_name).first()
+        if command.isdigit():
+            remove_index = int(command) - 1
+            item = session.query(Item).filter(Item.item_list == list_name).order_by(Item.id).all()[remove_index]
+        else:
+            item = session.query(Item).filter(Item.item_list == list_name, Item.item_name == command).first()
+        if not item:
+            raise ValueError
         session.delete(item)
         session.commit()
         session.close()
-        removed_message += '"' + command + '" is eraf!'
-    except (AttributeError, ValueError, TypeError) as e:
+        removed_message += '"' + item.item_name + '" is eraf!'
+    except (AttributeError, TypeError) as e:
         logging.exception(e)
         removed_message += 'Ben je dom ofzo?'
         pass
     except IndexError:
         removed_message += 'Dit nummer staat niet op de lijst mongol'
+        pass
+    except ValueError:
+        removed_message += 'Dit item staat niet op de lijst mongol'
         pass
     context.bot.send_message(chat_id=update.effective_chat.id, text=removed_message)
